Removing atomix dependency completely away from suuchi-core and into a separate module
