+++ pom.xml
@@ -12,6 +12,7 @@
         <module>suuchi-examples</module>
         <module>suuchi-rocksdb</module>
         <module>suuchi-membership-scalecube</module>
+        <module>suuchi-membership-atomix</module>
     </modules>
 
     <url>https://github.com/ashwanthkumar/suuchi</url>
