+++ suuchi-membership-atomix/README.md
@@ -1,0 +1,6 @@
+# [Atomix](http://atomix.io/atomix/) based Suuchi Membership
+
+This module provides Membership implementation using Raft consensus algorithm.
+
+## License
+https://www.apache.org/licenses/LICENSE-2.0
